// frontend/client/types.gen.ts
export type PropertyData = {
    /**
     * List of accessibility types
     */
    accessibilities?: Array<string>;
    /**
     * Number of bathrooms
     */
    bathrooms: number;
    /**
     * Number of bedrooms
     */
    bedrooms: number;
    /**
     * Business name
     */
    businessName?: string | null;
    /**
     * City
     */
    city: string;
    /**
     * Property ID
     */
    id: number;
    /**
     * Image URL
     */
    image_url: string;
    /**
     * Property name
     */
    name: string;
    /**
     * Price
     */
    price: number;
    /**
     * Size in square feet
     */
    size_sqft: number;
    /**
     * Street address
     */
    street: string;
};